+++ api/providers/ContactProvider.cs
@@ -1,3 +1,4 @@
+using System.Data;
 using DTO.Contact;
 using FluentValidation;
 
@@ -24,6 +25,7 @@ public class ContactProvider : BaseProvider<Contact>
         Contact newContact = new Contact(newInstance: true)
         {
             Name = req.Name,
+            Function = req.Function,
             Phone = req.Phone,
             Email = req.Email
         };
@@ -41,6 +43,13 @@ public class ContactProvider : BaseProvider<Contact>
         Contact? foundContact = _db.Contacts.FirstOrDefault(c => c.Id == id);
         if (foundContact == null) return null;
 
+
+        if (_db.Warehouses.Any(w => w.ContactId == id) ||
+            _db.Clients.Any(c => c.ContactId == id) ||
+            _db.Suppliers.Any(s => s.ContactId == id))
+        {
+            throw new ApiFlowException($"{id} The provided contact_id is in use and cannot be modified or deleted.", StatusCodes.Status409Conflict);
+        }
         _db.Contacts.Remove(foundContact);
         SaveToDBOrFail();
 
@@ -53,9 +62,10 @@ public class ContactProvider : BaseProvider<Contact>
         if (req == null) throw new ApiFlowException("Invalid contact request. Could not update contact.");
 
         Contact? existingContact = _db.Contacts.FirstOrDefault(c => c.Id == id);
-        if (existingContact == null) throw new ApiFlowException($"Contact not found for id '{id}'");
+        if (existingContact == null) throw new ApiFlowException($"Contact not found for id '{id}'", StatusCodes.Status404NotFound);
 
         existingContact.Name = req.Name;
+        existingContact.Function = req.Function;
         existingContact.Phone = req.Phone;
         existingContact.Email = req.Email;
         existingContact.SetUpdatedAt();
@@ -69,16 +79,4 @@ public class ContactProvider : BaseProvider<Contact>
     }
 
     protected override void ValidateModel(Contact model) => _contactValidator.ValidateAndThrow(model);
-
-    // Andere methoden zoals GetOrCreateContact blijven zoals ze zijn
-    public Contact? GetOrCreateContact(ContactRequest? contact = null, Guid? contactId = null)
-    {
-        if (contact == null && contactId == null) return null;
-
-        if (contactId != null) return GetById(contactId.Value);
-
-        if (contact != null) return Create(contact);
-
-        return null;
-    }
 }
